Remove a deprecated copy of pulumi-ts-provider from tests
