--- tests/integration/packageadd-remote/provider/pulumi-ts-provider/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "pulumi-ts-provider",
-  "version": "0.1.0",
-  "private": true,
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsc",
-    "prepare": "npm run build"
-  },
-  "devDependencies": {
-    "@types/node": "^18.0.0",
-    "typescript": "^5.0.0"
-  },
-  "dependencies": {
-    "@pulumi/pulumi": "^3.145.0"
-  }
-}
