--- tests/integration/packageadd-remote/provider/pulumi-ts-provider/src/index.ts
@@ -1,1 +1,0 @@
-export {componentProviderHost} from './provider';
