--- tests/integration/packageadd-remote/provider/pulumi-ts-provider/tsconfig.json
@@ -1,17 +1,0 @@
-{
-    "compilerOptions": {
-      "target": "es2018",
-      "module": "commonjs",
-      "declaration": true,
-      "declarationMap": true,
-      "sourceMap": true,
-      "outDir": "./dist",
-      "strict": true,
-      "esModuleInterop": true,
-      "skipLibCheck": true,
-      "forceConsistentCasingInFileNames": true
-    },
-    "include": ["src"],
-    "exclude": ["node_modules", "dist", "**/*.test.ts"]
-  }
-
