+++ tests/integration/packageadd-remote/provider/ts/package.json
@@ -3,17 +3,11 @@
     "description": "Self Signed Certificate",
     "version": "0.1.3",
     "dependencies": {
-        "@pulumi/pulumi": "^3.144.0",
-        "@pulumi/tls": "^4.6.1",
-        "pulumi-ts-provider": "file:../pulumi-ts-provider/dist"
+        "@pulumi/pulumi": "^3.160.0",
+        "@pulumi/tls": "^4.6.1"
     },
     "devDependencies": {
         "@types/node": "^10.0.0",
-        "pkg": "^5.6.0",
         "typescript": "^4.0.0"
-    },
-    "scripts": {
-        "gen": "ts-node index.ts --gen",
-        "start": "ts-node index.ts"
     }
 }
